Resolves #31 Added Helper classes.
